+++ CHANGELOG.md
@@ -106,6 +106,8 @@
 - Added `\Wizaplace\SDK\Pim\OptionStatus`
 - Added `Wizaplace\SDK\Catalog\CatalogService::getDeclinationById`
 - Updated `Wizaplace\SDK\Catalog\Declination`
+- Updated `Wizaplace\SDK\Catalog\DeclinationOption`
+- Updated `Wizaplace\SDK\Catalog\Option`
 
 ## 1.42.0